+++ src/ro/webdata/qoli/aggr/stats/dimensions/gov/GovRightsPaths.java
@@ -11,6 +11,7 @@ public class GovRightsPaths {
     public static final String GENDER_EMP_GAP_FILE_NAME = "genderEmpGap";
     public static final String GENDER_PAY_GAP_FILE_NAME = "genderPayGap";
     public static final String POPULATION_TRUST_FILE_NAME = "populationTrustRatio";
+    public static final String POPULATION_TRUST_OTHER_FILE_NAME = "populationTrustOtherRatio";
     public static final String VOTER_TURNOUT_FILE_NAME = "voterTurnout";
 
     public static final String GOV_RIGHTS_RAW_PATH = String.join(File.separator, Constants.RAW_DATASET_PATH, GOVERNANCE_FILE_NAME);
@@ -24,5 +25,6 @@ public class GovRightsPaths {
             GENDER_EMP_GAP_PATH = generatePath(GENDER_EMP_GAP_FILE_NAME),
             GENDER_PAY_GAP_PATH = generatePath(GENDER_PAY_GAP_FILE_NAME),
             POPULATION_TRUST_PATH = generatePath(POPULATION_TRUST_FILE_NAME),
+            POPULATION_TRUST_OTHER_PATH = generatePath(POPULATION_TRUST_OTHER_FILE_NAME),
             VOTER_TURNOUT_PATH = GOV_RIGHTS_RAW_PATH + File.separator + VOTER_TURNOUT_FILE_NAME + Constants.CSV_EXTENSION;
 }